# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "DevEmailItem": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "to": {
            "title": "To",
            "type": "string"
          }
        },
        "required": [
          "to",
          "subject",
          "body",
          "event_id"
        ],
        "title": "DevEmailItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "senha": {
            "title": "Senha",
            "type": "string"
          }
        },
        "required": [
          "email",
          "senha"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "exp": {
            "format": "date-time",
            "title": "Exp",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "usuario": {
            "$ref": "#/components/schemas/UsuarioResumo"
          }
        },
        "required": [
          "token",
          "exp",
          "usuario"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
      "MeResponse": {
        "properties": {
          "exp": {
            "format": "date-time",
            "title": "Exp",
            "type": "string"
          },
          "usuario": {
            "$ref": "#/components/schemas/UsuarioResumo"
          }
        },
        "required": [
          "usuario",
          "exp"
        ],
        "title": "MeResponse",
        "type": "object"
      },
      "MensagemResponse": {
        "properties": {
          "mensagem": {
            "title": "Mensagem",
            "type": "string"
          }
        },
        "required": [
          "mensagem"
        ],
        "title": "MensagemResponse",
        "type": "object"
      },
...
      "PrimeiroAcessoRequest": {
        "properties": {
          "senha": {
            "title": "Senha",
            "type": "string"
          }
        },
        "required": [
          "senha"
        ],
        "title": "PrimeiroAcessoRequest",
        "type": "object"
      },
      "RecuperarSenhaRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "RecuperarSenhaRequest",
        "type": "object"
      },
      "RedefinirSenhaRequest": {
        "properties": {
          "senha": {
            "title": "Senha",
            "type": "string"
          }
        },
        "required": [
          "senha"
        ],
        "title": "RedefinirSenhaRequest",
        "type": "object"
      },
...
      "TrocarSenhaRequest": {
        "properties": {
          "nova_senha": {
            "minLength": 1,
            "title": "Nova Senha",
            "type": "string"
          },
          "senha_atual": {
            "title": "Senha Atual",
            "type": "string"
          }
        },
        "required": [
          "senha_atual",
          "nova_senha"
        ],
        "title": "TrocarSenhaRequest",
        "type": "object"
      },
...
      "UsuarioResumo": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "perfil": {
            "title": "Perfil",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "perfil",
          "status"
        ],
        "title": "UsuarioResumo",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SETES.DOCS API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout_auth_logout_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "me_auth_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/primeiro-acesso/{token}": {
      "post": {
        "description": "PRD US 1.6 Cen.1 — senha aceita, usuário é autenticado e status vira Ativo.",
        "operationId": "primeiro_acesso_auth_primeiro_acesso__token__post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrimeiroAcessoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Primeiro Acesso",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/recuperar-senha": {
      "post": {
        "description": "PRD US 1.3 Cen.3/Cen.5 — nunca revela se o e-mail está cadastrado.",
        "operationId": "recuperar_senha_auth_recuperar_senha_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecuperarSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recuperar Senha",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/redefinir-senha/{token}": {
      "post": {
        "description": "Zera bloqueio/tentativas na mesma transação (US 1.3 Cen.2c — desbloqueio automático).",
        "operationId": "redefinir_senha_auth_redefinir_senha__token__post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedefinirSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Redefinir Senha",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/trocar-senha": {
      "post": {
        "description": "PRD US 1.7 — exige senha atual correta; impede reuso das últimas 6 senhas.",
        "operationId": "trocar_senha_auth_trocar_senha_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrocarSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trocar Senha",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/internal/dev/emails": {
      "get": {
        "operationId": "listar_emails_internal_dev_emails_get",
        "parameters": [
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DevEmailItem"
                  },
                  "title": "Response Listar Emails Internal Dev Emails Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar Emails",
        "tags": [
          "dev"
        ]
      }
    },
...
  }
}